bases/communication-service/requirements.txt
Django==6.0
djangorestframework==3.16.1
psycopg[binary]==3.2.4
python-dotenv==1.2.1
celery==5.6.2
django-filter==25.2

django_filter==25.2
drf-spectacular==0.29.0


PyJWT==2.11.0

pytest==9.0.2
pytest-django==4.12.0

weasyprint==68.1
resend==2.23.0
isort==8.0.0
flake8==7.3.0
pylint==4.0.5
autopep8==2.3.2

